requirements.txt
streamlit
google-cloud-core
google-api-core
google-cloud-aiplatform
